+++ src/webview/components/modals/index.ts
@@ -2,3 +2,4 @@ export { Modal } from "./components/Modal";
 export { LinkModal } from "./components/LinkModal";
 export { ImageModal } from "./components/ImageModal";
 export { TableModal } from "./components/TableModal";
+export { SettingsModal } from "./components/SettingsModal";